CSS 작업 진행 중 (간단하게)

+++ WebContent/prod_trade/prod_trade_detail.jsp
@@ -6,47 +6,45 @@
 <%@page import="com.prod.db.ProdDTO"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
+
 <!DOCTYPE html>
 <html>
 <head>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-<!-- 합쳐지고 최소화된 최신 CSS -->
+ <!-- 합쳐지고 최소화된 최신 CSS --> 
 <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css">
-
-<!-- 부가적인 테마 -->
+ <!-- 부가적인 테마 --> 
 <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap-theme.min.css">
-
-<!-- 합쳐지고 최소화된 최신 자바스크립트 -->
+ <!-- 합쳐지고 최소화된 최신 자바스크립트 -->
 <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.2/js/bootstrap.min.js"></script>
 
 
 <meta charset="UTF-8">
-<title>Insert title here</title>
-<script src="./jq/jquery-3.6.0.js"></script>
+<!-- <script src="./jq/jquery-3.6.0.js"></script> -->
+<%@ include file="../inc/top.jsp" %>
 
 </head>
-<body>
-	<h1>WebContent/prod_trade/prod_trade_detail.jsp</h1>
 
+<body>
+	<!-- <h1>WebContent/prod_trade/prod_trade_detail.jsp</h1> -->
 	<%
 	ProdDTO pDTO = (ProdDTO) request.getAttribute("product");
 	String nick = (String)session.getAttribute("user_nick");
 	WishDTO wDTO = new WishDTO();
-	
-	
 	WishDAO wDAO = new WishDAO();
+	UserDAO uDAO = new UserDAO(); 
 	int pageNum = Integer.parseInt(request.getParameter("pageNum").toString());
-	
 	//int wishCount = (int)request.getAttribute("wishCount");
-	
 	%>
+	<!-- <a href="./main.bo">메인</a> -->
 
-	<a href="./main.bo">메인</a>
-	
-	<!-- 신고폼 -->
-	
-	<form action="./declarationProd.decl" method="post" onsubmit="return confirm('이 글을 신고하시겠습니까?')">
-		
+	<div class="container" >
+	<br><br>
+
+	<!-- 신고폼 -->	
+	<% if(user_nick != null){ %>
+	<div style="margin:auto;  width: 800px; ">
+	<form name="declareForm" action="./declarationProd.decl" method="post" onsubmit="return confirm('이 글을 신고하시겠습니까?')">
 		<input type="submit" value="신고하기">
 		<input type="hidden" name="prod_num" value="<%=pDTO.getProd_num()%>">
 		<!-- 신고 당하는 글 작성자 -->
@@ -54,17 +52,13 @@
 		<input type="hidden" name="board_sub" value="<%=pDTO.getProd_sub()%>">
 		<input type="hidden" name="board_type" value="1">
 		<input type="hidden" name="pageNum" value="<%=pageNum%>">
-	
 	</form>
+	</div>
+	<%} %>
+<!-- 	<form action="#" method="post" name="pfr"> -->
+<%-- 		<input type="hidden" name="nick" value=<%=nick%>> --%>
 	
-	<!-- 신고폼 -->
-	
-	
-	
-	<form action="#" method="post" name="pfr">
-	<input type="hidden" name="nick" value=<%=nick%>>
-	
-		<table border="1">
+		<table border="1" style="margin:auto;  width: 800px;">
 			<tr>
 				<td width="400">
 					 <!-- <img src="./upload/" width="400" height="400"> -->
@@ -81,19 +75,26 @@
 						<!-- Wrapper for slides -->
 						<div class="carousel-inner" role="listbox">
 							<div class="item active">
-								<img src="./upload/<%=pDTO.getProd_img().split(",")[0]%>" class="d-block w-100" alt="..."
+							
+							<%
+							String img0 = pDTO.getProd_img().split(",")[0] ; 
+							if ( img0 == null || img0.equals("null")){
+								img0 = "product_default.jpg";
+							}
+							%>
+								<img src="./upload/<%=img0%>" class="d-block w-100" alt="..."
 									style="width: 400px; height: 400px;"> 
 									
 								<div class="carousel-caption">...</div>
 							</div>
 			
 							<% 
-					
-							for (int i = 1; i < pDTO.getProd_img().split(",").length ; i++) {
-								String imgfile = pDTO.getProd_img().split(",")[i];
-								if ((imgfile == null) || (imgfile.equals("null"))) {
-									imgfile = "product_default.jpg";
-								}
+							if(pDTO.getProd_img().split(",").length > 1){
+								for (int i = 1; i < pDTO.getProd_img().split(",").length ; i++) {
+									String imgfile = pDTO.getProd_img().split(",")[i];
+									if ((imgfile == null) || (imgfile.equals("null"))) {
+										imgfile = "product_default.jpg";
+									}
 							%>
 							
 							<div class="item">
@@ -101,7 +102,10 @@
 									style="width: 400px; height: 400px;">
 								<div class="carousel-caption">...</div>
 							</div>
-							<%} %>
+							<%
+								} //for
+							} //if
+							%>
 						</div>
 
 						<!-- Controls -->
@@ -183,13 +187,16 @@
 					 case 2:
 					 	status = "무료나눔";
 					 	break;
+					 case 3:
+						 status ="판매완료"; 
+						 break;
 					 }
 					 %>
-										<ul>
+					<ul>
 						<li>카테고리 : <%=category%></li>
 						<li>거래여부 : <%=status%></li>
 						<li>조회수 : <%=pDTO.getProd_count() == 0 ? 1 : pDTO.getProd_count()%></li>
-						<small><li>작성시간 : <%=pDTO.getProd_date()%></li></small>
+						<li>작성시간 : <%=pDTO.getProd_date()%></li>
 					</ul> 
 					
 					찜 횟수 &nbsp; 
@@ -202,58 +209,44 @@
 					</c:if>
 					<span id="wish__Count"> <%=wDAO.wishCount(pDTO.getProd_num()) %> </span>
 					
-					
 					<!-- 관리자만 사용가능한 메뉴 생성 -->
-	
-					
-					<input type="button" id="btnLike" value="찜하기" class="form-control"> 
-					<input type="button" value="구매하기" class="form-control"> 
-					<input type="button" value="채팅하기" class="form-control">
-					
-
+					<% if(user_nick != null){ %>
+						<input type="button" id="btnLike" value="찜하기" class="form-control"> 
+						<input type="button" value="구매하기" class="form-control"> 
+						<input type="button" value="채팅하기" class="form-control">
+					<%}%>
 				</td>
 			</tr>
-			<tr>
-				<td colspan="2" height="400">
+			<tr style="border: 1px solid">
+				<td colspan="2" height="400" style="vertical-align: top">
 					<h1>상세정보</h1> <%=pDTO.getProd_content()%>
-
 				</td>
 			</tr>
-
 		</table>
-
-		<input type="button" value="수정하기"
-			onclick="location.href='./ProductModify.pr?num=<%=pDTO.getProd_num()%>'">
-
-		<input type="button" value="삭제하기"
-			onclick="location.href='./ProductDeleteAction.pr?num=<%=pDTO.getProd_num()%>'">
-
-
-	</form>
-
-
-
+		<div style="margin:auto;  width: 800px; ">
+		<% if ( pDTO.getUser_nick().equals(user_nick) ){ %>
+			<input type="button" value="수정하기"
+				onclick="location.href='./ProductModify.pr?num=<%=pDTO.getProd_num()%>'">
+		<%} %>
+		<% if( pDTO.getUser_nick().equals(user_nick) || uDAO.isAdmin(user_nick)){   %>
+			<input type="button" value="삭제하기"
+				onclick="location.href='./ProductDeleteAction.pr?num=<%=pDTO.getProd_num()%>'">
+		<%} %>
+		</div>
+<!-- 	</form> -->
 <!--  		if(user_nick == null){
 				alert("로그인 후 이용 가능합니다.");
 				location.href='./UserLogin.us' 
 			} -->
-			
-
-
-
+	</div>
 </body>
 
 <script type="text/javascript">
 	$(document).ready(function(){
-		
 		var user_nick = document.getElementById("nick");
-			
 		$(".hide").hide();
 		$("#btnLike").click(function(){
-			
-			
 			$.ajax({
-				
 				url:"favoriteProdAction.fp",
 				data:{
 					prod_num: <%=pDTO.getProd_num()%>,
@@ -262,7 +255,6 @@
 				type:"post",
 				datatype: "json",
 				success : function(data){
-					
 					// 찜목록 버튼 눌렀을 때 동작 
 					// 기본에 있으니까 빼고 하트 비우는 거.
 					if(data.check == 1){
@@ -277,21 +269,10 @@
 						$("#wish__Count").html(data.count);
 					}
 				}
-
-					
-				
 			})
-			
-			
 		});
-		
-		
 	});
-
-
-
-
 </script>
 
-
 </html>
+<%@ include file="../inc/footer.jsp" %>
